RdioObject: coalesce all extras for all types
For the ‘get’ call, we can’t really restrict extras to a specific
type since arbitrary keys (of arbitrary types) can be provided as
input.

Therefore, just coalesce all possible extras for all possible types
for passing to ’get’, and let it be super fucking slow on the server,
because the server is going to die. Because Rdio is going to die :(

Now I get to use this API and not give a shit about performance.